server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-genai>=1.33.0",
    "google-generativeai>=0.8.5",
    "openai>=1.105.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
